fix floating point number integration between database and api

+++ sql/schema/004_workouts_exercises.sql
@@ -2,7 +2,7 @@
 CREATE TABLE workouts_exercises (
     id UUID PRIMARY KEY,
     time_seconds INTEGER NOT NULL,
-    weight_kg INTEGER NOT NULL,
+    weight_kg REAL NOT NULL,
     workout_id UUID NOT NULL REFERENCES workouts (id) ON DELETE CASCADE,
     exercise_id UUID NOT NULL REFERENCES exercises (id) ON DELETE CASCADE
 );
