+++ sql/schema/005_rounds.sql
@@ -2,7 +2,7 @@
 CREATE TABLE rounds (
     id UUID PRIMARY KEY,
     round_number INT NOT NULL,
-    reps_completed DECIMAL NOT NULL,
+    reps_completed REAL NOT NULL,
     workout_exercise_id UUID REFERENCES workouts_exercises (id) ON DELETE CASCADE
 );
