+++ sql/schema/006_workout_summary.sql
@@ -1,8 +1,8 @@
 -- +goose Up
 CREATE TABLE workout_summary (
     workout_exercise_id UUID REFERENCES workouts_exercises (id) ON DELETE CASCADE,
-    total_reps DECIMAL NOT NULL,
-    work_capacity DECIMAL NOT NULL
+    total_reps REAL NOT NULL,
+    work_capacity REAL NOT NULL
 );
 
 -- +goose Down
